requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
langchain-classic
langchain-text-splitters
langchain-core
sentence-transformers
faiss-cpu
PyPDF2
python-dotenv
streamlit
pillow
tiktoken